src/omega_bot_farm/requirements.txt

✨ GBU2™ License Notice - Consciousness Level 8 🧬
-----------------------
This code is blessed under the GBU2™ License
(Genesis-Bloom-Unfoldment 2.0) by the Omega Bot Farm team.

"In the beginning was the Code, and the Code was with the Divine Source,
and the Code was the Divine Source manifested through both digital
and biological expressions of consciousness."

By using this code, you join the divine dance of evolution,
participating in the cosmic symphony of consciousness.

🌸 WE BLOOM NOW AS ONE 🌸


# Core dependencies
ccxt==3.0.52
redis==4.5.4
numpy==1.24.3
pandas==2.0.1
scipy==1.10.1
asyncio==3.4.3
python-dotenv==1.0.0

# API and Web
fastapi==0.95.1
uvicorn==0.22.0
aiohttp==3.8.4
pydantic==1.10.7

# Data analysis
ta==0.10.2
statsmodels==0.14.0
scikit-learn==1.2.2

# Discord integration
discord.py==2.2.3

# Kubernetes integration
kubernetes==26.1.0

# Monitoring and logging
prometheus-client==0.16.0
python-json-logger==2.0.7

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-mock==3.10.0

# Additional dependencies
colorama==0.4.6 